ml-service/requirements.txt
# Core ML Service Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0

# Image Processing
opencv-python==4.8.1.78
Pillow==10.1.0
numpy==1.24.3

# OCR Engines
pytesseract==0.3.10
paddlepaddle==2.5.2
paddleocr==2.7.0.3
easyocr==1.7.0

# Machine Learning
scikit-learn==1.3.2
sentence-transformers==2.2.2
torch==2.1.1
transformers==4.35.2

# OpenAI Integration
openai==1.3.7

# Utilities
python-dotenv==1.0.0
requests==2.31.0
aiofiles==23.2.1

# Development and Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
